+++ src/MEMZone.WriteLog/Public/Format-Message.ps1
@@ -57,11 +57,11 @@ function Format-Message {
 .LINK
     https://MEM.Zone
 .LINK
-    https://MEMZ.one/PSWriteLog
+    https://MEMZ.one/WriteLog
 .LINK
-    https://MEMZ.one/PSWriteLog-GIT
+    https://MEMZ.one/WriteLog-GIT
 .LINK
-    https://MEMZ.one/PSWriteLog-ISSUES
+    https://MEMZ.one/WriteLog-ISSUES
 .COMPONENT
     Script Utilities
 .FUNCTIONALITY