+++ src/MEMZone.WriteLog/Public/Invoke-WithAnimation.ps1
@@ -44,11 +44,11 @@ function Invoke-WithAnimation {
 .LINK
     https://MEM.Zone
 .LINK
-    https://MEMZ.one/PSWriteLog
+    https://MEMZ.one/WriteLog
 .LINK
-    https://MEMZ.one/PSWriteLog-GIT
+    https://MEMZ.one/WriteLog-GIT
 .LINK
-    https://MEMZ.one/PSWriteLog-ISSUES
+    https://MEMZ.one/WriteLog-ISSUES
 .COMPONENT
     Script Utilities
 .FUNCTIONALITY