+++ src/MEMZone.WriteLog/Public/Write-FunctionHeaderOrFooter.ps1
@@ -27,11 +27,11 @@ function Write-FunctionHeaderOrFooter {
 .LINK
     https://MEM.Zone
 .LINK
-    https://MEMZ.one/PSWriteLog
+    https://MEMZ.one/WriteLog
 .LINK
-    https://MEMZ.one/PSWriteLog-GIT
+    https://MEMZ.one/WriteLog-GIT
 .LINK
-    https://MEMZ.one/PSWriteLog-ISSUES
+    https://MEMZ.one/WriteLog-ISSUES
 .COMPONENT
     Script Utilities
 .FUNCTIONALITY